- Refactor code  - To make use of index.jsx syntax
