+++ src/components/bodyComponents/inventory/index.jsx
